Refactored writers into separate file
