+++ report.go
@@ -1,10 +1,7 @@
 package report
 
 import (
-	"encoding/json"
 	"errors"
-	"io"
-	"os"
 	"sync"
 	"time"
 
@@ -24,27 +21,6 @@ type Logger struct {
 
 // Data is a string-keyed map of unstructured data relevant to the event
 type Data map[string]interface{}
-
-// Writer is a event writing function
-type Writer func(d Data) error
-
-// And adds another writer to execute in parallel
-func (w Writer) And(next Writer) Writer {
-	return func(d Data) error {
-		ch := make(chan error)
-		go func() {
-			ch <- w(d)
-		}()
-		go func() {
-			ch <- next(d)
-		}()
-		if err := <-ch; err != nil {
-			<-ch
-			return err
-		}
-		return <-ch
-	}
-}
 
 type command int
 
@@ -77,33 +53,6 @@ func New(w Writer, global Data) *Logger {
 	}
 	go logger.run()
 	return &logger
-}
-
-// JSON writes JSON formatted logs
-func JSON(w io.Writer) Writer {
-	encoder := json.NewEncoder(w)
-	return func(d Data) error {
-		return encoder.Encode(d)
-	}
-}
-
-// StdOutJSON writes logs to StdOut as JSON
-func StdOutJSON() Writer {
-	return JSON(os.Stdout)
-}
-
-// Honeycomb sends log events to HoneyComb
-func Honeycomb(key string, dataset string) Writer {
-	libhoney.Init(libhoney.Config{
-		WriteKey: key,
-		Dataset:  dataset,
-	})
-
-	return func(d Data) error {
-		ev := libhoney.NewEvent()
-		ev.Add(d)
-		return ev.Send()
-	}
 }
 
 // Info logs event that provide telemetry measures or context to any events requiring action.
